backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0

# Semantic search dependencies (updated versions)
sentence-transformers>=3.0.0
faiss-cpu>=1.7.4
numpy>=1.24.0
torch>=2.0.0
